Implementation, over update-rs-closures, of the propagation of the offset-mark.

+++ src/share/vm/lag1/lag1OopClosures.inline.hpp
@@ -64,7 +64,7 @@ LAG1ParMarkFollowerClosure::do_oop_work(T * p, uint32_t m)
   oop obj = oopDesc::load_decode_heap_oop(p);
   assert (_worker_id == _par_scan_state->queue_num(), "sanity");
   // install the offset mark in the oop p
-  if(!obj->has_allocr() && obj->cas_install_allocr(m)) {
+  if(!obj->has_allocr() && obj->cas_install_allocr((uintptr_t)m)) {
 #ifdef LAG1_DEBUG_TRACING
     gclog_or_tty->print_cr("[lag1-debug-tracing] children oop " INTPTR_FORMAT " is marked, mark is "INTPTR_FORMAT,
                            (intptr_t)obj, (intptr_t)obj->mark());
